Fix email validator
